+++ src/Routes.js
@@ -1,0 +1,18 @@
+import React, { Component } from 'react';
+import {Router, Stack, Scene} from 'react-native-router-flux';
+
+import Login from './screens/Login';
+import Signup from './screens/Signup';
+
+export default class Routes extends Component<{}> {
+	render() {
+		return(
+			<Router>
+			    <Stack key="root" hideNavBar={true}>
+			      <Scene key="login" component={Login} title="Login" initial={true}/>
+			      <Scene key="signup" component={Signup} title="Register"/>
+			    </Stack>
+			 </Router>
+			)
+	}
+}
